remove hidden-md class from demo element
the hidden-md class never gets removed/added in the demo, and gets hidden at different breakpoints that what the controls are doing.  It appears like the control doesn't work in the demo.  Remove the class and it behaves correctly with all the settings toggled and screen resized to the various breakpoints.